requirements.txt
requests==2.31.0
pandas==2.1.4
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0
dash==2.14.2
beautifulsoup4==4.12.2
lxml==4.9.3
python-dotenv==1.0.0
streamlit==1.28.1
yfinance==0.2.18
openpyxl==3.1.2
xlsxwriter==3.1.9
